requirements.txt
python-whois==0.7.1
requests==2.22.0
